chore: move colors and variants
