+++ src/index.ts
@@ -1,1 +1,5 @@
 export * from './components';
+
+export * from './themes';
+
+export * from './constants';
